abstract Diary type definition into types.ts
